GUI sketching cont'd.
- Sketched GUI widgets
- Added new stock fonts (dejavu)
- moved old default font (Courier New) to obsolete
- Added node-specific material modifier support for shader
- Changed many ulong's to size_t's (for 32-bit support)

+++ engine/render/batch.d
@@ -108,7 +108,7 @@ class Batch
     
     Node[] nodes;    
 
-    ulong length() { return nodes.length; }
+    size_t length() { return nodes.length; }
 
     Node add(Node node) { nodes ~= node; return node; }
     void clear() { nodes.length = 0; }
@@ -208,9 +208,9 @@ class BatchGroup
         foreach(batch; batches) batch.clear();
     }
     
-    ulong length()
+    size_t length()
     {
-        ulong l = 0;
+        size_t l = 0;
         foreach(batch; batches) l += batch.length();
         return l;
     }
